web/api/requirements.txt

# Auth dependencies
cryptography
python-jose
# Basic dependencies
gunicorn
motor
requests
httpcore==0.13.6
Jinja2==2.11.3
MarkupSafe==1.1.1
urllib3==1.26.3
python-multipart
Pillow
pybase64
click==8.1.3
fastapi==0.92.0
idna==3.4
pydantic==1.10.5
python-dotenv==0.21.1
PyYAML==6.0
sniffio==1.3.0
starlette==0.25.0
uvicorn==0.20.0
websockets==10.4
typing==3.7.4.3